+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]
 ### Added
 - Table Extension Support
+- Handle before block output
 
 ## [0.2.1]
 ### Fixed